backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorclass==2.2.2
coreapi==2.3.3
coreschema==0.0.4
channels
channels-redis
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-jsonfield==1.4.1
django-rest-swagger==2.2.0
django-solo==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
idna==3.4
import-export==0.2.67.dev6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pip-upgrader==1.4.15
pipreqs==0.4.11
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
srt==3.5.2
terminaltables==3.1.10
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
yarg==0.1.9
django-environ==0.4.5